tests/wpt/web-platform-tests/css/css-text-decor/text-decoration-thickness-initial.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" title="2.4 Text Decoration Width: the 'text-decoration-thickness' property"
      href="https://drafts.csswg.org/css-text-decor-4/#text-decoration-width-property" />
</head>
<body>
<script>
test(function() {
  assert_equals(getComputedStyle(document.body)["text-decoration-thickness"], "auto", "Must be set to value auto as initial value.");
}, "Initial value of text-decoration-thickness should be auto");
</script>
</body>
</html>
